+++ frontend/src/api/admin/stats.ts
@@ -16,3 +16,10 @@ export const countCaRequestedCerts = (uuid: string, caOrSsl: boolean) =>
     pathNames: { uuid },
     searchParams: { caOrSsl }
   });
+
+export const countAllUsrs = (role: 1 | 2 | 3 | 0 = 0) =>
+  callRestfulApi<number>({
+    method: "GET",
+    baseUrl: "/api/v1/superadmin/user/count",
+    searchParams: { role }
+  });
