+++ frontend/src/api/index.ts
@@ -2,6 +2,8 @@ import type { ResultVO } from "@/api/types";
 import { useUserStore } from "@/stores/user";
 import router from "@/router";
 
+export let noTimeout = false;
+
 export const createURLSearchParams = (
   params: Record<string, any>
 ): URLSearchParams => {
@@ -35,7 +37,7 @@ export const callRestfulApi = async <U = null>(
         ? { "Content-Type": "application/json" }
         : undefined,
     body: opts.payload !== undefined ? JSON.stringify(opts.payload) : undefined,
-    signal: AbortSignal.timeout(10000)
+    signal: noTimeout ? undefined : AbortSignal.timeout(10000)
   } satisfies RequestInit;
 
   const resp = await fetch(uri, req);
